Create TodoContainer and modify code in App component based on final project rubric
